+++ indicatorppadownloadstatistics/src/indicatorppadownloadstatistics/CHANGELOG.md
@@ -3,9 +3,11 @@
 ## v1.0.81 (2024-09-04)
 
 - Reinstated the autostart option in Preferences with the addition of a optional delay to start up.
-- Added plucky to the list of series.
-- Added oracular to the list of series.
-- Removed the preference to combine PPAs; the algorithm breaks down for binary packages differing in version number and/or mixed architecture.
+- Removed the series/architecture for a given PPA (only need specify a PPA user/name).
+- Filters now apply to a PPA user/name, irrespecitve of the series/architecture.
+- Removed the preference to combine PPAs; instead, binary packages with the same name and version are combined.
+  - Where binary packages of the same name/version are architecture independant, the download count is the same across all packages.
+  - Where binary packages of the same name/version are architecture dependant, the download count is summed across all packages.
 - Now includes a symbolic icon allowing the colour to be adjusted for the current theme.
 - Overhauled to adhere to the pyproject.toml standard.
 - Deployment using the .deb format is superceded; PyPI (pip) is used with operating system packages and file copy.